+++ forge/src/main/java/pro/mikey/accelerateddecay/forge/AcceleratedDecayForge.java
@@ -1,14 +1,32 @@
 package pro.mikey.accelerateddecay.forge;
 
-import dev.architectury.platform.forge.EventBuses;
+import net.minecraft.server.level.ServerLevel;
+import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.world.level.Level;
+import net.minecraftforge.common.MinecraftForge;
+import net.minecraftforge.event.TickEvent;
+import net.minecraftforge.event.world.BlockEvent;
+import net.minecraftforge.eventbus.api.SubscribeEvent;
+import net.minecraftforge.fml.LogicalSide;
 import pro.mikey.accelerateddecay.AcceleratedDecay;
 import net.minecraftforge.fml.common.Mod;
-import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
-
 @Mod(AcceleratedDecay.MOD_ID)
 public class AcceleratedDecayForge {
     public AcceleratedDecayForge() {
-        EventBuses.registerModEventBus(AcceleratedDecay.MOD_ID, FMLJavaModLoadingContext.get().getModEventBus());
         AcceleratedDecay.init();
+
+        MinecraftForge.EVENT_BUS.register(this);
+    }
+
+    @SubscribeEvent
+    void onBreakEvent(BlockEvent.BreakEvent event) {
+        AcceleratedDecay.breakEvent((Level) event.getWorld(), event.getState(), event.getPos(), (ServerPlayer) event.getPlayer());
+    }
+
+    @SubscribeEvent
+    void onServerLevelTick(TickEvent.WorldTickEvent event) {
+        if (event.side == LogicalSide.SERVER && event.phase == TickEvent.Phase.END) {
+            AcceleratedDecay.levelTickPostEvent((ServerLevel) event.world);
+        }
     }
 }